Rename omniauth_credential folder to omniauthCredential
