+++ app/packs/entrypoints/application.js
@@ -8,7 +8,7 @@ var ChemScanner = require('src/apps/chemscanner/');
 var ChemSpectra = require('src/apps/chemspectra/ChemSpectra');
 var ChemSpectraEditor = require('src/apps/chemspectra/ChemSpectraEditor');
 var MoleculeModerator = require('src/apps/molecule_moderator');
-var OmniauthCredential = require('src/apps/omniauth_credential');
+var OmniauthCredential = require('src/apps/omniauthCredential');
 var UserCounter = require('src/apps/user_counter');
 var ScifinderCredential = require('src/apps/scifinder_credential');
 var mydb = require('src/apps/mydb');
