+++ app/packs/src/apps/omniauthCredential/OmniauthCredential.js
