+++ app/packs/src/apps/omniauthCredential/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import OmniauthCredential from 'src/apps/omniauth_credential/OmniauthCredential';
+import OmniauthCredential from 'src/apps/omniauthCredential/OmniauthCredential';
 
 document.addEventListener('DOMContentLoaded', () => {
   const domElement = document.getElementById('OmniauthCredential');
